% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_script.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map FARS data reports for a state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer or character vector for the state FIPS number to map.}

\item{year}{Integer or character vector for the year to map.}
}
\value{
map plot showing latitude and longitude locations of traffic fatalities in the specified
state and year. If the function is not able to coerce the inputs to integers the function will
generate an error. If the files are not found at the current working directory the function
will return an error. Issues reading in the CSV files also return errors.
}
\description{
This function creates a map of fatal vehicle traffic crashes in a specified state and year.
}
\examples{
fars_map_state(42, 2014)
fars_read_years(12, "2013")

}
