% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_script.R
\name{fars_read}
\alias{fars_read}
\title{Load single Fatality Analysis Reporting System (FARS) file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A file path where the function should look for the FARS data file. .gz, .bz2,
.xz, or .zip files will be automatically uncompressed. Files starting with http://, https://,
ftp://, or ftps:// will be automatically downloaded.}
}
\value{
tibble containing the specified FARS data. If the file is not found at the specified path
   the function will return an error. Issues reading in the CSV file also return errors.
}
\description{
This function reads a specified file as a tibble format (tbl_df class). The \code{filename}
argument should be the path of the file to read.
}
\examples{
fars_read("data.csv")
fars_read("accident_2013.csv.bz2")

}
