% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_script.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Load FARS data files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Integer or character vector for the years to load data for.}
}
\value{
separate tibbles for each year. If the function is not able to coerce the inputs
   to integers the function will generate an error. If the files are not found at the current working
   directory the function will return an error. Issues reading in the CSV files also return errors.
}
\description{
This function loads FARS files for specified years.
}
\examples{
fars_read_years(c(2014,2015))
fars_read_years("2013")

}
