% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_script.R
\name{make_filename}
\alias{make_filename}
\title{Create file name for FARS data file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Integer or character string for the year to find data for.}
}
\value{
file path for the specified annual FARS data. If the function is not able to coerce the input
   to an integer the function will generate an error.
}
\description{
This function creates a pathname f
or FARS data for each specified year.
}
\examples{
make_filename(2015)
make_filename("2013")

}
